Fix all the bugs related to keeping pointers to Process
They're in a dynamic datastructure, so that's a nono.

We could choose to manage the ownership more manually and store a
BTreeMap of <usize, NonNull<Process>> or something, but this does work
for now, even if there are minor innefficiencies (like copying Context a
few extra times).

+++ kernel/src/pmm.rs
@@ -126,9 +126,9 @@ pub fn alloc_contiguous(pages: usize) -> Option<u64> {
     None
 }
 
-pub fn free(page: usize) {
+pub fn free(page: u64) {
     let mut page_info = PAGE_INFO.lock();
-    let page = page / 4096;
+    let page = (page / 4096) as usize;
     match page_info[page] {
         PageInfo::InUse { refcount } => {
             if refcount == 1 {
